Leaving a chat closes only its panel, keeping the chat list (#7561)
(cherry picked from commit cbf59f9153b4bb985c5b2d66a0cf5978f3190c42)
