+++ test/pangea/navigation/activity_room_close_test.dart
@@ -1,19 +1,16 @@
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:fluffychat/features/navigation/panel_types_enum.dart';
+import 'package:fluffychat/features/navigation/room_close_location.dart';
 import 'package:fluffychat/features/navigation/route_facts.dart';
-import 'package:fluffychat/routes/chat/activity_sessions/activity_sessions_start_view.dart';
 
 void main() {
   Uri u(String s) => Uri.parse(s);
 
-  group('activityRoomCloseLocation (#7156)', () {
+  group('roomTokenCloseLocation (#7156, #7561)', () {
     test('closing a room-token activity drops only the room, keeping the '
         'chat list', () {
-      final loc = activityRoomCloseLocation(
-        u('/?left=chats,room:!abc'),
-        '!abc',
-      );
+      final loc = roomTokenCloseLocation(u('/?left=chats,room:!abc'), '!abc');
       expect(loc, isNotNull);
       final left = parseOpenPanels(u(loc!)).left;
       expect(
@@ -27,7 +24,7 @@ void main() {
     });
 
     test('other open panels survive (e.g. an analytics summary)', () {
-      final loc = activityRoomCloseLocation(
+      final loc = roomTokenCloseLocation(
         u('/?left=chats,room:!abc&right=analytics:vocab'),
         '!abc',
       );
@@ -39,9 +36,23 @@ void main() {
       );
     });
 
+    test('leaving a chat opened from the chat list keeps the list and the '
+        'course context (#7561)', () {
+      final loc = roomTokenCloseLocation(
+        u('/?c=!course&left=chats,room:!abc'),
+        '!abc',
+      );
+      expect(loc, isNotNull);
+      final closed = u(loc!);
+      expect(parseOpenPanels(closed).left.map((t) => t.type), [
+        PanelTypesEnum.chats,
+      ]);
+      expect(activeSpaceIdFor(closed), '!course'); // map scope survives
+    });
+
     test('the standalone /<activityId> route has no room token, so it returns '
         'null (caller pops or falls back to home)', () {
-      expect(activityRoomCloseLocation(u('/abc123'), '!abc'), isNull);
+      expect(roomTokenCloseLocation(u('/abc123'), '!abc'), isNull);
     });
   });
 }
